Known issues:
- Clock analysis pass under printf synthesis has different behavior
- Async Reset seems to break FAME5 transform
- Autocounter nees to be ported

+++ sim/src/main/scala/midasexamples/Parity.scala
@@ -10,7 +10,7 @@ class ParityDUT extends Module {
     val in  = Input(Bool())
     val out = Output(Bool())
   })
-  val s_even :: s_odd :: Nil = Enum(UInt(), 2)
+  val s_even :: s_odd :: Nil = Enum(2)
   val state  = RegInit(s_even)
   when (io.in) {
     when (state === s_even) { state := s_odd  }
